Refactor axis options menu for reuse
