+++ src-js/fontra-core/src/location-tools.js
@@ -1,3 +1,5 @@
+import { translate } from "@fontra/core/localization.js";
+
 /*
  * Automatic location space conversion for settings objects and other location logic
  * that is shared between the font overview and the glyph editor.
@@ -91,4 +93,70 @@ export function filterLocation(location, axes) {
   }
 
   return filteredLocation;
+}
+
+export const ShowLocationSettings = Object.freeze({
+  DontShowEffectiveLocation: 0,
+  ShowEffectiveLocation: 1,
+  OnlyShowEffectiveLocation: 2,
+});
+
+export function getAxisOptionsMenuItems(settings, forHiddenAxes) {
+  const effectiveLocationKey = forHiddenAxes
+    ? "hiddenFontAxesShowEffectiveLocation"
+    : "fontAxesShowEffectiveLocation";
+  const menuItems = [
+    {
+      title: translate(
+        "sidebar.designspace-navigation.font-axes-view-options-menu.apply-single-axis-mapping"
+      ),
+      callback: () => {
+        settings.fontAxesUseSourceCoordinates = !settings.fontAxesUseSourceCoordinates;
+      },
+      checked: !settings.fontAxesUseSourceCoordinates,
+    },
+    {
+      title: translate(
+        "sidebar.designspace-navigation.font-axes-view-options-menu.apply-cross-axis-mapping"
+      ),
+      callback: () => {
+        settings.fontAxesSkipMapping = !settings.fontAxesSkipMapping;
+      },
+      checked: !settings.fontAxesSkipMapping,
+    },
+    { title: "-" },
+    {
+      title: translate(
+        "sidebar.designspace-navigation.font-axes-view-options-menu.show-effective-location"
+      ),
+      callback: () => {
+        settings[effectiveLocationKey] =
+          settings[effectiveLocationKey] == ShowLocationSettings.ShowEffectiveLocation
+            ? ShowLocationSettings.DontShowEffectiveLocation
+            : ShowLocationSettings.ShowEffectiveLocation;
+      },
+      checked:
+        settings[effectiveLocationKey] == ShowLocationSettings.ShowEffectiveLocation,
+    },
+  ];
+
+  if (forHiddenAxes) {
+    menuItems.push({
+      title: translate(
+        "sidebar.designspace-navigation.font-axes-view-options-menu.show-only-effective-location"
+      ),
+      callback: () => {
+        settings[effectiveLocationKey] =
+          settings[effectiveLocationKey] ==
+          ShowLocationSettings.OnlyShowEffectiveLocation
+            ? ShowLocationSettings.DontShowEffectiveLocation
+            : ShowLocationSettings.OnlyShowEffectiveLocation;
+      },
+      checked:
+        settings[effectiveLocationKey] ==
+        ShowLocationSettings.OnlyShowEffectiveLocation,
+    });
+  }
+
+  return menuItems;
 }
